Split in-memory runtime modules
